+++ remote_config.go
@@ -30,6 +30,9 @@ const (
 // Path to the local config for dumping/loading.
 const configPath = "config.json"
 
+// Path to the local config for dumping/loading if `configPath` is not writable.
+const tmpConfigPath = "/tmp/gobrake_remote_config.json"
+
 type remoteConfig struct {
 	opt *NotifierOptions
 	// opt copy to capture the initial state of the local config.
@@ -69,9 +72,17 @@ func (rc *remoteConfig) Poll() {
 	rc.pollStop = make(chan bool)
 
 	go func() {
-		if err := loadConfig(rc.JSON); err == nil {
+		// Load config from the standard location. If that fails, load
+		// from /tmp.
+		err := loadConfig(configPath, rc.JSON)
+		if err != nil {
+			if err = loadConfig(tmpConfigPath, rc.JSON); err == nil {
+				rc.updateLocalConfig()
+			}
+		} else {
 			rc.updateLocalConfig()
 		}
+
 		if err := rc.tick(); err != nil {
 			logger.Print(err)
 		}
@@ -147,8 +158,10 @@ func (rc *remoteConfig) StopPolling() {
 		rc.pollStop <- true
 	}
 
-	if err := dumpConfig(rc.JSON); err != nil {
-		logger.Printf("dumpConfig failed: %s", err)
+	if err := dumpConfig(configPath, rc.JSON); err != nil {
+		if err := dumpConfig(tmpConfigPath, rc.JSON); err != nil {
+			logger.Printf("dumpConfig failed: %s", err)
+		}
 	}
 }
 
@@ -257,13 +270,13 @@ func buildRequest(url string) (*http.Request, error) {
 	return req, nil
 }
 
-func dumpConfig(j *RemoteConfigJSON) error {
+func dumpConfig(path string, j *RemoteConfigJSON) error {
 	b, err := json.Marshal(j)
 	if err != nil {
 		return err
 	}
 
-	f, err := os.OpenFile(configPath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0666)
+	f, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0666)
 	if err != nil {
 		return err
 	}
@@ -280,8 +293,12 @@ func dumpConfig(j *RemoteConfigJSON) error {
 	return nil
 }
 
-func loadConfig(j *RemoteConfigJSON) error {
-	f, _ := ioutil.ReadFile(configPath)
+func loadConfig(path string, j *RemoteConfigJSON) error {
+	f, err := ioutil.ReadFile(path)
+	if err != nil {
+		return err
+	}
+
 	if err := json.Unmarshal(f, &j); err != nil {
 		return err
 	}